+++ lumin/data_processing/file_proc.py
@@ -1,24 +1,48 @@
 import h5py
-from h5py import Group
 import numpy as np
 import pandas as pd
-from typing import List, Union, Optional
+from typing import List, Union, Optional, Any
 import os
 from pathlib import Path
 
 from sklearn.model_selection import StratifiedKFold, KFold
 
 
-def save_to_grp(arr:np.ndarray, grp:Group, name:str) -> None:
-    '''Save array group in HDF5 file'''
+def save_to_grp(arr:np.ndarray, grp:h5py.Group, name:str) -> None:
+    r'''
+    Save Numpy array as a dataset in an h5py Group
+    
+    Arguments:
+        arr: array to be saved
+        grp: group in which to save arr
+        name: name of dataset to create
+    '''
+
+    # TODO Option for string lenght
+
     ds = grp.create_dataset(name, shape=arr.shape, dtype=arr.dtype.name if arr.dtype.name != 'object' else 'S16')
     ds[...] = arr if arr.dtype.name != 'object' else arr.astype('S16')
 
 
 def fold2foldfile(df:pd.DataFrame, out_file:h5py.File, fold_idx:int,
-                  cont_feats:List[str], cat_feats:List[str], targ_feats:Union[str,List[str]], targ_type:str,
+                  cont_feats:List[str], cat_feats:List[str], targ_feats:Union[str,List[str]], targ_type:Any,
                   misc_feats:Optional[List[str]]=None, wgt_feat:Optional[str]=None) -> None:
-    '''Save fold data into foldfile group'''
+    r'''
+    Save fold of data into an h5py Group
+    Arguments:
+        df: Dataframe from which to save data
+        out_file: h5py file to save data in
+        fold_idx: ID for the fold; used name h5py group according to `'fold_{fold_idx}'`
+        cont_feats: list of columns in df to save as continuous variables
+        cat_feats: list of columns in df to save as discreet variables
+        targ_feats (list of) column(s) in df to save as target feature(s)
+        targ_type: type of target feature, e.g. `int`,`'float32'`
+        misc_feats (optional): any extra columns to save
+        wgt_feat (optional): column to save as data weights
+    '''
+
+    # TODO infer target type automatically
+
     grp = out_file.create_group(f'fold_{fold_idx}')
     save_to_grp(np.hstack((df[cont_feats].values.astype('float32'), df[cat_feats].values.astype('float32'))), grp, 'inputs')
     save_to_grp(df[targ_feats].values.astype(targ_type), grp, 'targets')
@@ -29,8 +53,22 @@ def fold2foldfile(df:pd.DataFrame, out_file:h5py.File, fold_idx:int,
 
 def df2foldfile(df:pd.DataFrame, n_folds:int, cont_feats:List[str], cat_feats:List[str],
                 targ_feats:Union[str,List[str]], savename:Union[Path,str], targ_type:str,
-                strat_key:str=None, misc_feats:Optional[List[str]]=None, wgt_feat:Optional[str]=None):
-    '''Convert dataframe into foldfile'''
+                strat_key:Optional[str]=None, misc_feats:Optional[List[str]]=None, wgt_feat:Optional[str]=None):
+    r'''
+    Convert dataframe into h5py file by splitting data into sub-folds to be accessed by a :class:`FoldYielder`
+    Arguments:
+        df: Dataframe from which to save data
+        n_folds: number of folds to split df into
+        cont_feats: list of columns in df to save as continuous variables
+        cat_feats: list of columns in df to save as discreet variables
+        targ_feats (list of) column(s) in df to save as target feature(s)
+        savename: name of h5py file to create (`.h5py` extension not required)
+        targ_type: type of target feature, e.g. `int`,`'float32'`
+        strat_key (optional): column to use for stratified splitting
+        misc_feats (optional): any extra columns to save
+        wgt_feat (optional): column to save as data weights
+    '''
+
     savename = str(savename)
     os.system(f'rm {savename}.hdf5')
     os.makedirs(savename[:savename.rfind('/')], exist_ok=True)